requirements.txt
langchain
openai
unstructured
faiss-cpu
gradio
sentence_transformers
huggingface_hub
python-dotenv